seedvr2/pytorch/requirements.txt
omegaconf
rotary-embedding-torch
